A merge of shinki's changes from post #235. It seems like changes to Studio_Items.1.regex.txt were based off an older version than master,  where the file was possibly empty. A cursory examination does not reveal any conflicts, so I simply appended the file on top of master's.
Thread Quote for shinki-235:

Here's a translation for studio mods. This is my first attempt at translating, so there might be a few mistakes.

It includes translations for:

- New liquid items here (http://www.hongfire.com/forum/forum/hentai-lair/hf-modding-translation/honey-select-mods/5771534)
- Shame ♀ play ! - ASoul's Toy Shop here (http://www.hongfire.com/forum/forum/hentai-lair/hf-modding-translation/honey-select-mods/5771905)
- YashiroAmamiya Mod thread here (http://www.hongfire.com/forum/forum/hentai-lair/hf-modding-translation/honey-select-mods/5721026)
- Elysium mod thread here (http://www.hongfire.com/forum/forum/hentai-lair/hf-modding-translation/honey-select-mods/5748264)
- Stuff from http://erogeteikoku.blog.fc2.com/

Place the txt file in the Plugins\UITranslation\HoneyStudio\Text folder and hopefully it'll work.

Most work but some don't work properly (like the skydomes, new liquid, shame play items and a couple of weapons). By this I mean that they are translated in the menu where you select them from, but when you put them into the scene they display the Japanese version up the top in the scene bar. I'm not sure why that happens, I've tried to fix it a few times but I don't fully understand how this works so I'm at a loss.

If anyone can help it'll be greatly appreciated.

7/03/2017: File updated to include the missing mod items mentioned by tmvaegis. Thanks for the help btw.
8/03/2017: Added studio mods from http://hilite000.blog.fc2.com as requested by Super Hentai Man. The File is now split into two parts, regex and noregex. I sure hope this all works.

http://www.hongfire.com/forum/forum/hentai-lair/hf-modding-translation/honey-select-mods/5716891-wip-all-in-one-english-translation?p=5772756#post5772756