Clean up
After the format was organized a few of the files had the wrong
order/value for the headings, thus the position was slightly off.
Updated all headings to be the same across all files, and re-ordered
them into the order the correct files were. Technically c10 is still
missing the ADV section, but that’s a duplicate section anyway and
already resides within the Dialogue .txt file. Maybe one day I’ll move
it.