Merging in scharmers' 270
Scharmers' Good Enough Translations for HoneySelect

I've been going over the translation files in the latest RC. Some of
them are obviously hand-written, some are all just GT dumps. With the
amount of lines, this is to be expected! My task is to go over and
hand-rewrite the ones that need some love. I fixed content,
punctuation, and grammar to the best of my ability.

These files will overwrite the RC ones, or any other ones of the same
name. Since they may be more than one replacement source out there, you
may want to scan through my translations first, to see if they meet
your needs.

All mad props go to the original translators, who put in the hard work
to actually dump the lines for us along with the rest of their work.

Dialogue_c02.16.txt is for the gentle personality.